refactor: move withDB to test helper package
